Refactor ElasticBase to expose framework elements
I want to reuse Schema for a new API outside of the Elasticsearch
query context. I've pulled the conversion/normalization framework
up a level to server.api.resources, and reworked the post/put/get
delegation to (hopefully) support more general interfaces. (I expect
to use GET, PUT, DELETE, and possibly POST for metadata.)
