+++ lib/pbench/server/api/resources/query_apis/datasets_publish.py
@@ -5,16 +5,17 @@ from logging import Logger
 from flask_restful import abort
 
 from pbench.server import PbenchServerConfig
-from pbench.server.api.resources.query_apis import (
+from pbench.server.api.resources import (
     API_OPERATION,
     CONTEXT,
-    ElasticBase,
     JSON,
-    PostprocessError,
     Schema,
     Parameter,
     ParamType,
+    PostprocessError,
 )
+from pbench.server.api.resources.query_apis import ElasticBase
+
 from pbench.server.database.models.datasets import Dataset, Metadata
 from pbench.server.database.models.users import User
