+++ lib/pbench/server/api/resources/query_apis/elasticsearch_api.py
@@ -1,14 +1,8 @@
 from logging import Logger
 
 from pbench.server import PbenchServerConfig
-from pbench.server.api.resources.query_apis import (
-    CONTEXT,
-    ElasticBase,
-    JSON,
-    Schema,
-    Parameter,
-    ParamType,
-)
+from pbench.server.api.resources import CONTEXT, JSON, Schema, Parameter, ParamType
+from pbench.server.api.resources.query_apis import ElasticBase
 
 
 class Elasticsearch(ElasticBase):
